reformulando services

+++ Back End/src/main/java/com/wr/jonstory/repositories/PagamentoComBoletoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PagamentoComBoletoRepository extends JpaRepository<PagamentoComBoleto, Integer> {
+public interface PagamentoComBoletoRepository extends JpaRepository<PagamentoComBoleto, Long> {
 }
